# Add JMH benchmarks and apply optimizations