--- a/src/main/java/io/r2dbc/postgresql/message/backend/AuthenticationMD5Password.java
+++ b/src/main/java/io/r2dbc/postgresql/message/backend/AuthenticationMD5Password.java
@@ -18,6 +18,7 @@ package io.r2dbc.postgresql.message.backend;
 
 import io.netty.buffer.ByteBuf;
 import io.r2dbc.postgresql.util.Assert;
+import io.r2dbc.postgresql.util.ByteBufferUtils;
 
 import java.nio.ByteBuffer;
 import java.util.Objects;
@@ -38,7 +39,7 @@ public final class AuthenticationMD5Password implements AuthenticationMessage {
     public AuthenticationMD5Password(ByteBuf salt) {
         Assert.requireNonNull(salt, "salt must not be null");
 
-        this.salt = salt.nioBuffer();
+        this.salt = ByteBufferUtils.toByteBuffer(salt);
     }
 
     @Override