Before:

```
Benchmark                                 (resultSize)   Mode  Cnt      Score   Error  Units
StagedResultSizeBenchmarks.simpleJdbc                1  thrpt    5  13306,454 ± 1208,805  ops/s
StagedResultSizeBenchmarks.simpleJdbc               10  thrpt    5  10698,935 ± 1765,078  ops/s
StagedResultSizeBenchmarks.simpleJdbc              100  thrpt    5   8129,170 ± 2008,986  ops/s
StagedResultSizeBenchmarks.simpleJdbc              200  thrpt    5   5909,088 ± 1475,809  ops/s
StagedResultSizeBenchmarks.extendedJdbc              1  thrpt    5  23413,053 ± 2871,815  ops/s
StagedResultSizeBenchmarks.extendedJdbc             10  thrpt    5  20849,660 ±  999,644  ops/s
StagedResultSizeBenchmarks.extendedJdbc            100  thrpt    5   7913,647 ±  771,162  ops/s
StagedResultSizeBenchmarks.extendedJdbc            200  thrpt    5   3876,876 ±  842,278  ops/s
StagedResultSizeBenchmarks.simpleR2dbc               1  thrpt    5   9297,426 ±  627,729  ops/s
StagedResultSizeBenchmarks.simpleR2dbc              10  thrpt    5   6760,997 ±  327,497  ops/s
StagedResultSizeBenchmarks.simpleR2dbc             100  thrpt    5   2638,489 ±  287,265  ops/s
StagedResultSizeBenchmarks.simpleR2dbc             200  thrpt    5   1363,660 ±  136,011  ops/s
StagedResultSizeBenchmarks.extendedR2dbc             1  thrpt    5   7619,045 ±  563,956  ops/s
StagedResultSizeBenchmarks.extendedR2dbc            10  thrpt    5   6742,465 ±  380,895  ops/s
StagedResultSizeBenchmarks.extendedR2dbc           100  thrpt    5   2807,188 ±   90,361  ops/s
StagedResultSizeBenchmarks.extendedR2dbc           200  thrpt    5   1309,214 ±   41,054  ops/s
```

After:

```
Benchmark                                 (resultSize)   Mode  Cnt      Score      Error  Units
StagedResultSizeBenchmarks.simpleJdbc                1  thrpt    5  13306,454 ± 1208,805  ops/s
StagedResultSizeBenchmarks.simpleJdbc               10  thrpt    5  10698,935 ± 1765,078  ops/s
StagedResultSizeBenchmarks.simpleJdbc              100  thrpt    5   8129,170 ± 2008,986  ops/s
StagedResultSizeBenchmarks.simpleJdbc              200  thrpt    5   5909,088 ± 1475,809  ops/s
StagedResultSizeBenchmarks.extendedJdbc              1  thrpt    5  23413,053 ± 2871,815  ops/s
StagedResultSizeBenchmarks.extendedJdbc             10  thrpt    5  20849,660 ±  999,644  ops/s
StagedResultSizeBenchmarks.extendedJdbc            100  thrpt    5   7913,647 ±  771,162  ops/s
StagedResultSizeBenchmarks.extendedJdbc            200  thrpt    5   3876,876 ±  842,278  ops/s
StagedResultSizeBenchmarks.simpleR2dbc               1  thrpt    5  10713,126 ± 1283,533  ops/s
StagedResultSizeBenchmarks.simpleR2dbc              10  thrpt    5  10046,317 ±  784,692  ops/s
StagedResultSizeBenchmarks.simpleR2dbc             100  thrpt    5   5079,499 ± 1738,124  ops/s
StagedResultSizeBenchmarks.simpleR2dbc             200  thrpt    5   3875,689 ±  752,872  ops/s
StagedResultSizeBenchmarks.extendedR2dbc             1  thrpt    5  11612,125 ± 2114,247  ops/s
StagedResultSizeBenchmarks.extendedR2dbc            10  thrpt    5   9109,973 ± 1904,214  ops/s
StagedResultSizeBenchmarks.extendedR2dbc           100  thrpt    5   5314,664 ±  503,736  ops/s
StagedResultSizeBenchmarks.extendedR2dbc           200  thrpt    5   3507,015 ±  539,890  ops/s
```