@gregturn 2019-09-17
I think we need a separate polish commit that removes author tags.